% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine.R
\name{computeCosineDist}
\alias{computeCosineDist}
\title{Compute Cosine Distance}
\usage{
computeCosineDist(mat)
}
\arguments{
\item{mat}{matrix.\cr Cosine similarity matrix}
}
\value{
Return a cosine distance matrix.
}
\description{
Compute Cosine Distance
}
