% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine.R
\name{computeCosineSim}
\alias{computeCosineSim}
\title{Compute Cosine Similarity}
\usage{
computeCosineSim(mat)
}
\arguments{
\item{mat}{matrix.\cr A matrix combined candidate and reference sub-cluster
expression matrix.}
}
\value{
Return a cosine similarity matrix.
}
\description{
Compute Cosine Similarity
}
