% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{computeLassoScore}
\alias{computeLassoScore}
\title{Compute Lasso Score of BMMC in RNA-seq.}
\usage{
computeLassoScore(data, normalize = T)
}
\arguments{
\item{data}{matrix.\cr Candidate RNA-seq data in acute leukemia cells.}

\item{normalize}{logical.\cr Whether to normalize RNA-seq data by normalizeQuantiles.
Default is TRUE.}
}
\value{
Return lasso score of BMMC in RNA-seq.
}
\description{
Compute Lasso Score of BMMC in RNA-seq.
}
