% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{computeMeanLike}
\alias{computeMeanLike}
\title{Compute Mean of Like Value for Each Sub-clusters}
\usage{
computeMeanLike(hemato.subc, cell.subc, top.subc = 5)
}
\arguments{
\item{hemato.subc}{\code{\link{HematoSubCluster}}\cr HematoSubCluster object.}

\item{cell.subc}{character.\cr Names of sub-cluster}

\item{top.subc}{integer.\cr Number of top like sub-clusters for each candidate
sub-cluster. Default is 5.}
}
\value{
Return mean value of like value of each sub-clusters
}
\description{
Compute Mean of Like Value for Each Sub-clusters
}
