% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{computePredictSubClustePercentage}
\alias{computePredictSubClustePercentage}
\title{Compute Cell Percentage of Predicted Sub-clusters.}
\usage{
computePredictSubClustePercentage(like.mat, top.subc = 5, ref.subc)
}
\arguments{
\item{like.mat}{matrix.\cr Matrix of like.}

\item{top.subc}{integer.\cr Number of top like sub-clusters for each candidate
sub-cluster. Default is 5.}

\item{ref.subc}{character.\cr Names of reference sub-cluster.}
}
\value{
Return cell percentage of sub-clusters.
}
\description{
Compute Cell Percentage of Predicted Sub-clusters.
}
