% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{computeThetaLike}
\alias{computeThetaLike}
\title{Compute Cosine Theta and Corresponding Cell Like by Cosine Distance}
\usage{
computeThetaLike(hemato.subc, cell.subc = rownames(hemato.subc@meta.data))
}
\arguments{
\item{hemato.subc}{\code{\link{HematoSubCluster}}\cr HematoSubCluster object.}

\item{cell.subc}{character.\cr Names of Candidate sub-cluster}
}
\value{
Return a list containing a vector of cosine theta and a matrix of cell like.
}
\description{
Compute Cosine Theta and Corresponding Cell Like by Cosine Distance
}
