% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CosineDist-class}
\alias{CosineDist-class}
\alias{CosineDist}
\title{CosineDist-class}
\description{
A class for storing Cosine-Distance information.
}
\section{Slots}{

\describe{
\item{\code{sim}}{matrix.\cr Cosine similarity matrix.}

\item{\code{dist}}{matrix.\cr Cosine distance matrix.}

\item{\code{features.enroll}}{character.\cr Common features between input and reference data.}

\item{\code{subc.ref}}{character.\cr The name of reference sub-cluster. It can distinguish
between reference sub-cluster and candidate sub-cluster in cosine similarity matrix and
cosine distance matrix.}
}}

