% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/fun.R
\name{CreateSubclusterObject}
\alias{CreateSubclusterObject}
\alias{CreateSubclusterObject.Seurat}
\alias{CreateSubclusterObject.matrix}
\title{Create Sub-cluster Object of HematoSubCluster}
\usage{
CreateSubclusterObject(object, ...)

\method{CreateSubclusterObject}{Seurat}(
  object,
  assay = "RNA",
  reduction = "umap",
  cell.range = 20,
  cell.size = 50,
  set.seed = 1,
  n.core = 1,
  ...
)

\method{CreateSubclusterObject}{matrix}(object, meta.data, ...)
}
\arguments{
\item{object}{Seurat, matrix.\cr Seurat object or sub-cluster expression matrix.}

\item{...}{ANY.\cr Arguments passed to other methods.}

\item{assay}{character.\cr Assay of Seurat object used to calculate sub-cluster
expression matrix. Default is "RNA".}

\item{reduction}{character.\cr Reduction of Seurat object used to calculate
reduciont position of sub-cluster. Default is "umap".}

\item{cell.range}{integer.\cr Range of change of cells number in a sub-cluster
when runing Kmean. Default is 20.}

\item{cell.size}{integer.\cr Number of cells in a sub-cluster when runing Kmean.
Default is 50.}

\item{set.seed}{integer.\cr Seed number for Kmean to compute sub-cluster of cell cluster.
Default is 1.}

\item{n.core}{integer.\cr Number of core for parallel job of running MinMax Kmean cluster.
Default is 1.}

\item{meta.data}{data.frame.\cr Meta data information of sub-cluster.}
}
\value{
Return a object of \code{\link{HematoSubCluster}}, containing sub-cluster
expression matrix, and cosine similarity matrix.
}
\description{
Create Sub-cluster Object of HematoSubCluster
}
