% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{findDiffBetweenSubc}
\alias{findDiffBetweenSubc}
\title{Find Differentially Expressed Genes Between Sub-CellType}
\usage{
findDiffBetweenSubc(
  hemato.subc,
  seurat.obj,
  ref.subc = c("HSC/MPP", "LMPP", "CLP", "CMP", "MDP", "GMP", "CDP", "MEP", "GMP-Mono",
    "CD14 Mono", "CD16 Mono", "pre-DC", "pDC", "cDC1", "mo-DC", "MKP", "MK", "pro-Ery1",
    "pro-Ery2", "Ery", "pre-pro-B", "Early pro-B", "Early cycling pro-B", "Late pro-B",
    "Late cycling pro-B", "pre-B", "Immature B", "Naive B", "Memory B 1", "Memory B 2",
    "CD8 Tnaive", "CD8 Teff", "CD8 Tex", "CD8 Tdpe", "CD8 Tmpe", "CD4 Tnaive", "CD4 Tem",
    "CD4 Treg", "NK", "NK-XCL1", "cycling NK", "cycling NK/T"),
  target.subc = c("LMPP", "CLP", "CMP", "MDP", "GMP", "CDP", "MEP", "GMP-Mono",
    "CD14 Mono", "CD16 Mono", "pre-DC", "pDC", "cDC1", "mo-DC", "MKP", "MK", "pro-Ery1",
    "pro-Ery2", "Ery", "pre-pro-B", "Early pro-B", "Early cycling pro-B", "Late pro-B",
    "Late cycling pro-B", "pre-B", "Immature B", "Naive B", "Memory B 1", "Memory B 2",
    "CD8 Tnaive", "CD8 Teff", "CD8 Tex", "CD8 Tdpe", "CD8 Tmpe", "CD4 Tnaive", "CD4 Tem",
    "CD4 Treg", "NK", "NK-XCL1", "cycling NK", "cycling NK/T")
)
}
\arguments{
\item{hemato.subc}{\code{HematoSubCluster} object.\cr An object representing hematopoietic subclusters.}

\item{seurat.obj}{\code{Seurat} object.\cr A Seurat object containing single-cell RNA-seq data.}

\item{ref.subc}{character.\cr Reference subpopulation for comparison. Default is 'HSC/MPP'.}

\item{target.subc}{character.\cr Target subpopulation for comparison. Default is 'LMPP'.}
}
\value{
A data frame of differentially expressed genes between the reference and target sub-celltype.
}
\description{
Find Differentially Expressed Genes Between Sub-CellType
}
