% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{geom_bracket}
\alias{geom_bracket}
\title{Add Brackets with Labels to a GGPlot}
\usage{
geom_bracket(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  label = NULL,
  xmin = NULL,
  xmax = NULL,
  size = 0.3,
  label.size = 2.5,
  family = "",
  vjust = 0,
  coord.flip = FALSE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix}

\item{position}{Position adjustment, either as a string naming the adjustment,
or the result of a call to a position adjustment function. Use the latter
if you need to change the settings of the adjustment.}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with a
warning.  If \code{TRUE} silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification.}

\item{label}{Character vector with alternative label, if not null test is
ignored.}

\item{xmin}{Numeric vector with the positions of the left sides of the
brackets.}

\item{xmax}{Numeric vector with the positions of the right sides of the
brackets.}

\item{size}{The width of the lines of the bracket}

\item{label.size}{The size of the label text}

\item{family}{The font used for the text}

\item{vjust}{Move the text up or down relative to the bracket}

\item{coord.flip}{If \code{TRUE}, flip x and y coordinates so that
horizontal becomes vertical, and vertical, horizontal.}

\item{...}{other arguments passed on to \code{\link{layer}}.}
}
\description{
add brackets with label annotation to a ggplot. Helpers for
adding p-value or significance levels to a plot. \strong{It is based on \code{ggpubr::geom_bracket}}
}
