% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has.rownames}
\alias{has.rownames}
\title{Determine whether a Data.frame has Row Names}
\usage{
has.rownames(data)
}
\arguments{
\item{data}{data.frame.\cr A data frame.}
}
\value{
Return TRUE or FALSE for whether a data.frame has row names.
}
\description{
Determine whether a Data.frame has Row Names
}
