% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HematoSubCluster-class}
\alias{HematoSubCluster-class}
\alias{HematoSubCluster}
\title{HematoSubCluster-class}
\description{
A class for sub-cluster to store information, including expression matrix,
distance matrix, meta-info and other.
}
\section{Slots}{

\describe{
\item{\code{data}}{matrix.\cr Sub-cluster expression.}

\item{\code{dist}}{list.\cr Distance informations of sub-cluster.}

\item{\code{meta.data}}{data.frame.\cr Meta data information of sub-cluster.}

\item{\code{misc}}{list.\cr Utility slot for storing additional data.}
}}

