% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{message.time}
\alias{message.time}
\title{Print Current Time and Message}
\usage{
message.time(...)
}
\arguments{
\item{...}{ANY.\cr Message.}
}
\value{
Print string with time.
}
\description{
Print Current Time and Message
}
