% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalizeQuantiles}
\alias{normalizeQuantiles}
\title{Normalize Columns of a Matrix to have the same Quantiles}
\usage{
normalizeQuantiles(A, ties = TRUE)
}
\arguments{
\item{A}{numeric matrix.\cr Missing values are allowed.}

\item{ties}{logical.\cr If TRUE, ties in each column of A are treated in careful
way. tied values will be normalized to the mean of the corresponding pooled
quantiles.}
}
\value{
A matrix of the same dimensions as A containing the normalized values.
}
\description{
normalizeQuantiles function is a copy version from \strong{limma} package.
}
