% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCircleTreeRefFeature}
\alias{plotCircleTreeRefFeature}
\title{Plot Circle Tree with Feature Expression Mapping in Sub-cluster.}
\usage{
plotCircleTreeRefFeature(
  feature,
  scell.data,
  color.mapping = c("cell.type", "cell.value"),
  label.population = T,
  title = NULL
)
}
\arguments{
\item{feature}{character.\cr A feature in sub-cluster expression matrix.}

\item{scell.data}{matrix.\cr Sub-cluster expression matrix.}

\item{color.mapping}{character.\cr c("cell.type", "cell.value"). "cell.type"
means the color mapping of points is based on different cell types. "cell.value"
means the color mapping of points is based on the different feature expression
values of cell types.}

\item{label.population}{logical.\cr Whether to label population names. Default is TRUE.}

\item{title}{character.\cr Title of the plot. Default is NULL.}
}
\description{
Plot Circle Tree with Feature Expression Mapping in Sub-cluster.
}
