% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotClusterTree}
\alias{plotClusterTree}
\title{Plot Cluster Dendrogram Tree with Lineage Mapping in scRNA-seq.}
\usage{
plotClusterTree(
  hemato.subc,
  group.subc = c("external", "internal", "reference"),
  plot.scell = NULL,
  rm.nkt = T,
  top.subc = 5,
  color.mapping = c("cell.type", "cell.percentage"),
  size.mode = c("absolute.pct", "relative.pct"),
  relative.compute = c("ratio", "difference"),
  label.cell = F,
  label.population = T,
  label.pop.size = 3,
  line.alpha = 0.7,
  line.size = 0.35,
  line.size.ratio = 0.05,
  ribbon.size = 5,
  point.size = 10,
  title = NULL,
  text.size = 7
)
}
\arguments{
\item{hemato.subc}{\code{\link{HematoSubCluster}}\cr HematoSubCluster object.}

\item{group.subc}{charater.\cr c("external", "internal", "reference").
"external" means plot the whole sub-clusters and cell clusters of the
candidate sample. "internal" means plot the whole sub-clusters and cell
clusters of the reference in the same cosine similarity matrix. "reference"
means directly plot the reference sub-clusters and cell clusters info.}

\item{plot.scell}{character.\cr Candidate sub-cluster name. If provided, it will
only plot the like value of a sub-cluster to reference sub-cluster and cell cluster.}

\item{rm.nkt}{logical.\cr Weather to remove NK and T cell when plotting the like
value of a sub-cluster to reference sub-cluster and cell cluster.}

\item{top.subc}{integer.\cr Number of top like sub-clusters for each candidate
sub-cluster.}

\item{color.mapping}{character.\cr c("cell.type", "cell.percentage"). "cell.type"
means the color mapping of points is based on different cell types. "cell.percentage"
means the color mapping of points is based on the different percentages of cell types.}

\item{size.mode}{character.\cr c("absolute.pct", "relative.pct"). "absolute.pct"
means the size of points is based on actual percentages. "relative.pct" means
the size of points is based on percentages relative to normal BMMC.}

\item{relative.compute}{character.\cr c("ratio", "difference"). Method for calculate
relative to normal BMMC. It will only take effect when parameter size.mode selects
"relative.pct". "ratio" means the method is cell percentages of acute leukemia
divided by that of normal BMMC. "difference" means the method is cell percentages of
acute leukemia minus that of normal BMMC.}

\item{label.cell}{logical.\cr Whether to label cell names. Default is FALSE.}

\item{label.population}{logical.\cr Whether to label population names. Default is TRUE.}

\item{label.pop.size}{logical.\cr Text size of population names. Default is 3.}

\item{line.alpha}{numeric.\cr Alpha of line. Default is 0.7}

\item{line.size}{numeric.\cr Size of line. Default is 0.35.}

\item{line.size.ratio}{numeric.\cr Ratio of line size when line size is mapping.
Default is 0.05}

\item{ribbon.size}{numeric.\cr Size of lineage ribbon. Default is 5.}

\item{point.size}{numeric.\cr Size of points. Default is 10.}

\item{title}{character.\cr Title of the plot. Default is NULL.}

\item{text.size}{numeric.\cr Size of text. Default is 7.}
}
\value{
Return a plot of dendrogram tree with lineage info of acute leukemia cells.
}
\description{
Circle dendrogram tree is designed to show lineage info of acute
leukemia cells.
}
