% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotLassoLollipop}
\alias{plotLassoLollipop}
\title{Plot Lollipop with Lasso Score of Acute Leukemia Cells in RNA-seq.}
\usage{
plotLassoLollipop(mat.score, point.size = 3, line.size = 0.35, text.size = 7)
}
\arguments{
\item{mat.score}{matrix.\cr Lasso score matrix.}

\item{point.size}{numeric.\cr Size of points. Default is 3.}

\item{line.size}{numeric.\cr Size of line. Default is 0.35.}

\item{text.size}{numeric.\cr Size of text. Default is 7.}
}
\value{
Return a plot of lollipop with lasso score of acute leukemia cells.
}
\description{
Plot Lollipop with Lasso Score of Acute Leukemia Cells in RNA-seq.
}
