% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotLassoTree}
\alias{plotLassoTree}
\title{Plot Lasso Dendrogram Tree with Lineage Mapping of Acute Leukemia Cells in RNA-seq.}
\usage{
plotLassoTree(
  mat.score,
  color.mapping = c("value", "cell.type"),
  label.cell = F,
  label.population = T,
  label.pop.size = 3,
  line.alpha = 0.7,
  line.size = 0.35,
  line.size.ratio = 1.5,
  ribbon.size = 5,
  point.size = 1,
  point.color.limits = c(-2.2, 2.2),
  title = NULL,
  text.size = 7,
  legend.barheight = unit(2.5, "mm")
)
}
\arguments{
\item{mat.score}{matrix.\cr Lasso score matrix.}

\item{color.mapping}{character.\cr c("value", "cell.type"). "value" means
the color mapping of points is based on lasso score value. "cell.type" means
the color mapping of points is based on different cell types.}

\item{label.cell}{logical.\cr Whether to label cell names. Default is FALSE.}

\item{label.population}{logical.\cr Whether to label population names. Default is TRUE.}

\item{label.pop.size}{logical.\cr Text size of population names. Default is 3.}

\item{line.alpha}{numeric.\cr Alpha of line. Default is 0.7}

\item{line.size}{numeric.\cr Size of line. Default is 0.35.}

\item{line.size.ratio}{numeric.\cr Ratio of line size when line size is mapping.
Default is 1.5.}

\item{ribbon.size}{numeric.\cr Size of lineage ribbon. Default is 5.}

\item{point.size}{numeric.\cr Size of points. Default is 1.}

\item{point.color.limits}{numeric.\cr Limits of point color. Default is c(-2.2, 2.2)}

\item{title}{character.\cr Title of the plot. Default is NULL.}

\item{text.size}{numeric.\cr Size of text. Default is 7.}

\item{legend.barheight}{unit.\cr Bar height of legend. Default is unit(2.5, "mm").}
}
\value{
Return a plot of dendrogram tree with lasso score of acute leukemia cells.
}
\description{
Plot Lasso Dendrogram Tree with Lineage Mapping of Acute Leukemia Cells in RNA-seq.
}
