% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMeanLike}
\alias{plotMeanLike}
\title{Plot Different Mean of Like Value between Normal BMMC and Candidate Group.}
\usage{
plotMeanLike(
  hemato.subc,
  top.subc = 5,
  group.name = "Group",
  color = c("blue", "red"),
  label.size = 2.45,
  line.size = 0.35,
  text.size = 7
)
}
\arguments{
\item{hemato.subc}{\code{\link{HematoSubCluster}}\cr HematoSubCluster object.}

\item{top.subc}{integer.\cr Number of top like sub-clusters for each candidate
sub-cluster. Default is 5.}

\item{group.name}{character.\cr Name of candidate group. Default is "Group".}

\item{color}{character.\cr Color for two groups. Default is c("blue","red").}

\item{label.size}{numeric.\cr Size of label. Default is 2.45.}

\item{line.size}{numeric.\cr Size of line. Default is 0.35.}

\item{text.size}{numeric.\cr Size of text. Default is 7.}
}
\value{
Return a plot of different mean of like value between normal BMMC and
candidate group.
}
\description{
Plot Different Mean of Like Value between Normal BMMC and Candidate Group.
}
