% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotThetaLike}
\alias{plotThetaLike}
\title{Plot Cosine Theta and Like Value}
\usage{
plotThetaLike(
  hemato.subc,
  plot.scell = "C0:0002",
  point.size = 0.5,
  label.subc.number = 20,
  label.subc.size = 1,
  label.like.size = 2.5,
  line.size = 0.35,
  text.size = 7
)
}
\arguments{
\item{hemato.subc}{\code{\link{HematoSubCluster}}\cr HematoSubCluster object.}

\item{plot.scell}{character.\cr Candidate sub-cluster name. Default is "C0:0002".}

\item{point.size}{numeric.\cr Size of points. Default is 0.5.}

\item{label.subc.number}{integer.\cr Number of sub-cluster label. Default is 20.}

\item{label.subc.size}{numeric.\cr Size of sub-cluster label. Default is 1.}

\item{label.like.size}{numeric.\cr Size of like value. Default is 2.5.}

\item{line.size}{numeric.\cr Size of line. Default is 0.35.}

\item{text.size}{numeric.\cr Size of text. Default is 7.}
}
\value{
Return a plot of cosine theta and like value of a candidate sub-cluster.
}
\description{
Plot cosine theta and like value of a candidate sub-cluster to determine the
similarity of a candidate sub-cluster.
}
