% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove.subc.suffix}
\alias{remove.subc.suffix}
\title{Remove Sub-cluster Suffix}
\usage{
remove.subc.suffix(x, correct.cycling = F)
}
\arguments{
\item{x}{character.\cr Sub-cluster names with number suffix}

\item{correct.cycling}{logical.\cr}
}
\value{
Return a cosine distance matrix.
}
\description{
Remove Sub-cluster Suffix
}
