% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine.R
\name{rowNormalize}
\alias{rowNormalize}
\title{Normalize Matrix by Row}
\usage{
rowNormalize(m)
}
\arguments{
\item{m}{matrix.\cr A matrix for row normalization.}
}
\value{
Return a normalized matrix.
}
\description{
Normalize Matrix by Row
}
