% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runApp}
\alias{runApp}
\title{Run Shiny-based Interactive User Interface}
\usage{
runApp(options = list())
}
\arguments{
\item{options}{Named options that should be passed to the runApp call (these
can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify width and height
parameters which provide a hint to the embedding environment about the
ideal height/width for the app.}
}
\description{
Run Shiny-based Interactive User Interface
}
