% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/fun.R
\name{runCosineSimilarity}
\alias{runCosineSimilarity}
\alias{runCosineSimilarity.matrix}
\title{Run Cosine Similarity Analysis}
\usage{
runCosineSimilarity(object, ...)

\method{runCosineSimilarity}{matrix}(object, scell.ref)
}
\arguments{
\item{object}{ANY.\cr Input object.}

\item{...}{ANY.\cr Arguments passed to other methods.}

\item{scell.ref}{matrix.\cr Reference sub-cluster matrix used for evaluating
the similarity of candidate sub-cluster}
}
\value{
Return a object of \code{\link{CosineDist}}.
}
\description{
Run Cosine Similarity Analysis
}
