% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{runSeuratWorkflow}
\alias{runSeuratWorkflow}
\title{Run standard Seurat workflow}
\usage{
runSeuratWorkflow(object, assay = "RNA", force.run = F)
}
\arguments{
\item{object}{Suerat.\cr Seurat object.}

\item{assay}{character.\cr Assay name. Default is "RNA".}

\item{force.run}{logical.\cr Force to run standard Seurat workflow. Default is FALSE.}
}
\value{
Return Seurat object with normalizing, scaling, PCA, clustering.
}
\description{
Run standard Seurat workflow
}
