% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/fun.R
\name{runSubcluster}
\alias{runSubcluster}
\alias{runSubcluster.Seurat}
\title{Run MinMax-Kmean Cluster for Sub-cluster}
\usage{
runSubcluster(object, ...)

\method{runSubcluster}{Seurat}(
  object,
  cluster.key = "seurat_clusters",
  cell.range = 20,
  cell.size = 50,
  set.seed = 1,
  n.core = 1
)
}
\arguments{
\item{object}{ANY.\cr Input object.}

\item{...}{ANY.\cr Arguments passed to other methods.}

\item{cluster.key}{character.\cr The key of clustering data in Seurat object.
Default is "seurat_clusters".}

\item{cell.range}{integer.\cr Range of change of cells number in a sub-cluster
when runing Kmean. Default is 20.}

\item{cell.size}{integer.\cr Number of cells in a sub-cluster when runing Kmean.
Default is 50.}

\item{set.seed}{integer.\cr Seed number for Kmean to compute sub-cluster of cell cluster.
Default is 1.}

\item{n.core}{integer.\cr Number of core for parallel job of running MinMax Kmean cluster.
Default is 1.}
}
\value{
Return a vector containing sub-cluster info.
}
\description{
Run MinMax-Kmean Cluster for Sub-cluster
}
